Some moving and refactoring Add documentation page
