+++ packages/app/src/pages/actueel/gemeente/[code].tsx
@@ -35,7 +35,7 @@ import { useAgegroupLabels } from '~/domain/vaccine/logic/use-agegroup-labels';
 import { useIntl } from '~/intl';
 import { useFeature } from '~/lib/features';
 import { getWarning } from '~/queries/create-elements-query';
-import { getTopicalPageData } from '~/queries/topical-page-query';
+import { getTopicalPageData } from '~/queries/get-topical-page-data';
 import {
   createGetStaticProps,
   StaticProps,
