+++ packages/app/src/pages/gemeente/[code]/positief-geteste-mensen.tsx
@@ -23,10 +23,8 @@ import {
   getTimelineEvents,
 } from '~/queries/create-elements-query';
 import {
-  ArticleParts,
   getPagePartsQuery,
   isArticleParts,
-  PagePartQueryResult,
 } from '~/queries/get-page-parts.query';
 import {
   createGetStaticProps,
@@ -39,6 +37,7 @@ import {
   selectGmData,
 } from '~/static-props/get-data';
 import { filterByRegionMunicipalities } from '~/static-props/utils/filter-by-region-municipalities';
+import { ArticleParts, PagePartQueryResult } from '~/types/cms';
 import { replaceComponentsInText } from '~/utils/replace-components-in-text';
 import { replaceVariablesInText } from '~/utils/replace-variables-in-text';
 import { useReverseRouter } from '~/utils/use-reverse-router';
