+++ packages/app/src/pages/landelijk/positief-geteste-mensen.tsx
@@ -25,24 +25,23 @@ import { useIntl } from '~/intl';
 import {
   createElementsQuery,
   ElementsQueryResult,
-  getTimelineEvents,
+  getTimelineEvents
 } from '~/queries/create-elements-query';
 import {
-  ArticleParts,
   getPagePartsQuery,
-  isArticleParts,
-  PagePartQueryResult,
+  isArticleParts
 } from '~/queries/get-page-parts.query';
 import {
   createGetStaticProps,
-  StaticProps,
+  StaticProps
 } from '~/static-props/create-get-static-props';
 import {
   createGetChoroplethData,
   createGetContent,
   getLastGeneratedDate,
-  selectNlData,
+  selectNlData
 } from '~/static-props/get-data';
+import { ArticleParts, PagePartQueryResult } from '~/types/cms';
 import { replaceComponentsInText } from '~/utils/replace-components-in-text';
 import { useReverseRouter } from '~/utils/use-reverse-router';
