+++ packages/app/src/pages/landelijk/varianten.tsx
@@ -14,12 +14,9 @@ import { VariantsTableTile } from '~/domain/variants/variants-table-tile';
 import { useIntl } from '~/intl';
 import { withFeatureNotFoundPage } from '~/lib/features';
 import {
-  ArticleParts,
   getPagePartsQuery,
   isArticleParts,
   isLinkParts,
-  LinkParts,
-  PagePartQueryResult,
 } from '~/queries/get-page-parts.query';
 import {
   createGetStaticProps,
@@ -30,6 +27,7 @@ import {
   getLastGeneratedDate,
   selectNlData,
 } from '~/static-props/get-data';
+import { ArticleParts, LinkParts, PagePartQueryResult } from '~/types/cms';
 
 export const getStaticProps = withFeatureNotFoundPage(
   'nlVariantsPage',
