+++ packages/app/src/queries/get-page-parts.query.ts
@@ -1,64 +1,11 @@
-import { ArticleSummary } from '~/components/article-teaser';
-import { RichContentBlock } from '~/types/cms';
-
-export type PageIdentifier =
-  | 'in_positiveTestsPage'
-  | 'hospitalPage'
-  | 'in_positiveTestsPage'
-  | 'in_variantsPage'
-  | 'behaviorPage'
-  | 'situationsPage'
-  | 'reproductionPage'
-  | 'infectiousPeoplePage'
-  | 'topicalPage'
-  | 'elderlyAtHomePage'
-  | 'disabilityCarePage'
-  | 'positiveTestsPage'
-  | 'variantsPage'
-  | 'sewerPage'
-  | 'intensiveCarePage'
-  | 'vaccinationsPage'
-  | 'in_variantsPage'
-  | 'nursingHomePage'
-  | 'deceasedPage';
-
-export type PageBasePart = {
-  pageDataKind: string;
-};
-
-export type ArticleParts = {
-  _type: 'pageArticles';
-  articles: ArticleSummary[];
-} & PageBasePart;
-
-export type LinkParts = {
-  _type: 'pageLinks';
-  links: {
-    title: string;
-    href: string;
-  }[];
-} & PageBasePart;
-
-export type HighlightedItemParts = {
-  _type: 'pageHighlightedItems';
-  highlights: ArticleSummary[];
-  showWeeklyHighlight: boolean;
-} & PageBasePart;
-
-export type RichTextParts = {
-  _type: 'pageRichText';
-  text: RichContentBlock[];
-} & PageBasePart;
-
-export type PagePart =
-  | ArticleParts
-  | LinkParts
-  | HighlightedItemParts
-  | RichTextParts;
-
-export type PagePartQueryResult<T extends PagePart = PagePart> = {
-  pageParts: T[];
-};
+import {
+  ArticleParts,
+  HighlightedItemParts,
+  LinkParts,
+  PageIdentifier,
+  PagePart,
+  RichTextParts,
+} from '~/types/cms';
 
 export function isArticleParts(value: PagePart): value is ArticleParts {
   return value._type === 'pageArticles';
